feat: Filter sets that are likely 'rubber bands'
After filtering for sets that zoom to infinity, and sets that are periodic, a large number of sets are 'rubber bands' - a single, usually-convex loop. They're boring.

This commit adds a new test that detects whether a set is likely to be a loop of this sort (it also detects minorly-concave loops). Using the current settings, it has a very low false-positive rate. Its false-negative rate is pretty high, but it still manages to filter out a lot of sets. This test filters out a bit less than 50% of the sets that remain after the two previously-mentioned tests.